fix: FOVY conversion, BVH depth limit
GLTF FOVY is already in radians, since it was converted it cause the camera to be super zoomed in.
I tested the pathtracer with my FoxDen scene. It runs into memory crashes due to unsigned short overflows (65536), so I limited the BVH depth to 15 to prevent that boundary from being crossed.

+++ src/sceneStructs.h
@@ -26,8 +26,8 @@ struct BvhNode
 {
     glm::vec3 maxBounds;
     glm::vec3 minBounds;
-    unsigned short leftChild;
-    unsigned short rightChild;
+    int leftChild;
+    int rightChild;
     // In num floats, so offset / 3 = triangle count
     int trisOffset;
     // In num floats, so 3 * triangle count
